+++ changelog.md
@@ -1,4 +1,6 @@
 # Changelog
+- return null instead of throwing exception, when gb-player is unknown (expected behaviour)
+   - fixes #57
 
 ### v 2.2.3
 - stop players from accepting invitations in blocked worlds